Filesharing (bugfix): fixing syntax error in mysql query
git-svn-id: http://svn.contrexxlabs.com/contrexx/trunk@26116 5ce06e5d-2b28-0410-aa6a-b3cd9cf52d8a

+++ modules/filesharing/lib/FilesharingLib.class.php
@@ -303,7 +303,9 @@ CODE
             }
         }
         // delete all expired or not existing files
-        $objDatabase->Execute("DELETE FROM " . DBPREFIX . "module_filesharing WHERE `id` IN (" . implode(',', $arrToDelete) . ")");
+        if(!empty($arrToDelete)) {
+            $objDatabase->Execute("DELETE FROM " . DBPREFIX . "module_filesharing WHERE `id` IN (" . implode(',', $arrToDelete) . ")");
+        }
     }
 
     /**
